Crate evercrypt

Evercrypt & HACL Rust bindings

High-level evercrypt bindings crates.

⚠️ Note: This crate is still work in progress. Don’t use in production just yet.

Features

By default the Evercrypt crate includes the random feature that allows generating random values (keys, nonces, etc.). But this is not verified code and uses the rand crate. It can be disabled with --no-default-features. Please bring your own randomness if you want to be safe.

RustCrypto AES

Evecrypt currently implements AES only for x64 CPUs with a certain set of CPU instructions. To provide AES for other platforms the Evercrypt crate uses the RustCrypto AES implementation when using --features rust-crypto-aes.

Platforms

Currently only Linux x64 and MacOS are supported. Windows builds are on the To Do list and should be supported in future.

Platform	Supported

MacOS	✅

MacOS Arm64	✅

iOS	✅

iOS Simulator x86_64	❌

Linux x64	✅

Linux x86	✅

Windows x64	✅

Arm64 Linux	✅

Arm32 Linux	✅

Benchmarks

To run benchmarks use cargo bench.

Tests

All primitives are tested against the Wycheproof test vectors. They can be run with cargo test. This will also run automatically generated binding tests from bindgen.

Modules

aead

Authenticated Encryption with Associated Data (AEAD)

digest

Hashing

ecdh

ECDH

ed25519

Ed25519

hkdf

HKDF

hmac

HMAC

p256

prelude

Prelude for Evercrypt. Include this to get access to all the public functions of Evercrypt.

rand_util

Utilities that provide randomness. Note that this currently uses the rand crate and should be moved to a more secure alternative.

signature

x25519

x25519