Module evercrypt::signature

Enums

Error

Signature errors.

Mode

Supported signature schemes.

Functions

key_gen

Generate a new key pair for the given mode. The function throws an error for P256 keys if no valid key can be generated in a reasonable time.

sign

Sign a message msg with the secret key sk and the given signature scheme (mode). For ECDSA the hash algorithm and a nonce have to be passed in as well.

verify

Verify a signature. Depending on the Mode, a hash mode has to be passed in. The public key pk, signature, and message msg are passed in as byte slices.