Crate evercrypt

Modules

aead
Authenticated Encryption with Associated Data (AEAD)

digest
Hashing

ecdh
ECDH

ed25519
Ed25519

hkdf
HKDF

hmac
HMAC

p256

prelude
Prelude for Evercrypt. Include this to get access to all the public functions of Evercrypt.

rand_util
Utilities that provide randomness. Note that this currently uses the rand crate and should be moved to a more secure alternative.

signature

x25519
x25519